protobuf_comm: support self-describing (ID, type) messages
There is a new add_message() method that does not take component ID and
message type as parameter. For this to work messages must have a
sub-enum defined with the name CompType, which has at least the two
entries COMP_ID and MSG_TYPE, which must contain the required
information.