clips: support silencing of particular facts and rules
If rule watching is enabled, the very frequently occuring updates to the
time and rfid-input facts and the associated firing rules are cluttering
the output and make it hard to read anything. The new features allows to
name particular facts and rules in the configuration which should be
exempted from watching.