clips: consolidate and move machine processing code
Machine specification is now an own fact template. It covers inputs,
outputs, and processing times of machines. No more machine
initialization or global variables are required. Rather, we assert it as
a deffact. This also allows us to reduce the field machine processing
rules to one set for all give machine types, rather than one set for
each. Machine lights are no longer set directly in the processing rules.
Rather, a machine has two new slots for the actual and desired lights
and a rule takes care of asserting the proper states.