clips: add state and phase to game state
The state is no longer an unordered fact but subsumed by the gamestate
deftemplate. Add the phase slot. The state now is simply running,
stopped etc., the phase is exploration or production. Machine processing
is only enabled in RUNNING/PRODUCTION. This is also the default state
for now during testing.