shell: update visuals only from main thread
Receiving a message (in another thread in the client) could cause the
screen to get garbled if another update was running concurrently. Use
Boost ASIO io_service's dispatch mechanism to dispatch a call to the
main thread or updating the GUI. This in particular fixes issues when a
menu was shown.