clips: pucks change state only on end, invalid puck detection
Pucks remain in their original state while placed at a processing
machine. Only after processing is done the pucks are transformed to the
CONSUMED state. This is required to keep track of the placed puck types
at machines. Detect an invalid situation where a puck is moved from one
machine to another. This invalidates the puck and sets its state to
CONSUMED. This should not happen, but may if the referees do not play
close enough attention.