clips: implement automatic game phase switching
The normal game flow has been implemented. E.g. switching to RUNNING
state in the PRE_GAME phase will switch to the EXPLORATION phase. After
the exploration time is over switch automatically to the PRODUCTION
phase, and then to the POST_GAME phase (and PAUSED state). Re-initialize
game-time to zero on switching to PRODUCTION phase.