clips: improved beacon handling, detect lost hosts and wrong time
On the initial beacon we now check the time difference and warn if it is
above a certain threshold (3 sec atm). Do not use the robot's supplied
time for checking aliveness but only local times. Warn if hosts maybe
lost and again if they are definitely lost. In that case also erase them
from the working memory.