clips: correctly handle remote puck loading and removal
When ordered to load or remove a puck in or out of a machine, the
machine state must be adapted accordingly. Since this essentially
modifies the internal state without going through the expected sequence
of signals, we need to handle this explictly. But we envision that
referees might decide that a good is in a machine area in some
situations (e.g. a field referee tipped it off). Hence we need to make
the effort.