shell: always show machine placing menu
When a puck is added to a machine always show the menu whether the puck
should be placed under the RFID sensor or loaded into the machine. It is
more confusing if this is decided automatically and there is no way to
abort the action (not being able to load could be an indicator that the
wrong machine was chosen).