protobuf_comm: use protobuf descriptor pool for msg creation
If the message has not been explicitly registered protobuf_comm can now
query the static protobuf DescriptorPool to which all linked messages
are added. If the messages have the appropriate CompType information
they are added to the map for quicker instantiation next time. This
means that for now only incoming message types must be registered, but
outgoing will be registered automatically on first use.