refbox: use generic message typing facilities
Configure paths where to look for proto files and only use dynamic
message generation. For our message volume it is efficient enough, but
it allows us to be agnostic to supported messages types within the C++
part, and only require the appropriate message files to be available for
the CLIPS environment.