clips: replace system by game time for processing and points
Processing time was calculated based on system time. Not only required
that more matches on the (time) fact, but it also gave wrong processing
times when the game was paused. Handle all game related times in game
time now. Additionally do not store the system time in the (points) fact
anymore, it would again require matching on the (time) fact, which could
cause wrong activation sequences in some situations.