clips: robot maintenance support
The refbox can now track robot maintenance. In particular it tracks the
time in maintenance, the number of times a robot has been in
maintenance, and it enforces rules, e.g. the automatic disqualification
of a robot if not back in time.

There is a grace period of 15 sec after the maintenance time is over.
This is done for one because the referee might need a short time to
re-enter the robot if it is ready only in the last second. For another
we will need some time for the teams to adjust to the strict regime.