clips: add support for maintaining the SETUP phase
If a real game is played, i.e. a team name has been set, starting the
game goes into the SETUP phase and from there automatically transitions
to the EXPLORATION phase after five minutes. For a training game the old
behavior (going straight to the EXPLORATION phase) remains the same.