protobuf_clips: do not run agenda on events
Events might be triggered from within the CLIPS environment, e.g. when
explicitly disonnecting a client. This would lead to the successful
acquisition of the (recursive) mutex lock. That in turn would cause the
CLIPS environment's run() function to be called multiple times
simultaneously corrupting the CLIPS context. Therefore, do not run the
agenda from the communicator, but rely on the outside to invoke it.