clips: support synchronizing game over and overtime
If a master-slave game is played the refbox will pause the game at the
end of the production phase and wait for the other refbox to reach the
same point. Then the master makes the decision whether to play overtime
or not. Both refboxes then go into overtime or call the game over.