protobuf_comm: open combined IPv4/IPv6 sockets in server
On a system with IPv6 support resolving localhost might actually result
in ::1 (IPv6) instead of 127.0.0.1 (IPv4). This made even local
connections from shell to refbox impossible if the IPv6 address was
preferred (as newer Boost or resolver versions seem to do). Opening a
port for IPv6 automatically opens it for IPv4 as well, so we should be
good to just use that for now. The future, it happened...