msgs: add team to OrderInfo and MachineInfo
Orders and Machines as well as the compound list carry team names now.
The reason is that MachineInfo grows too big in bytes to send as UDP
packet. Therefore we want to split this into messages per team. To make
it easy to filter for relevant data, we set the team color in the
compound message then. For the TCP protocol we can easily handle the
larger sizes and would prefer this for better efficiency. Then we set
the team color only in the Machine or Order and put them all in a single
compound message.