clips: remove all known pucks when game starts
If not in training mode (i.e. we are using the PLC field) we erase all
pucks after randomization. Otherwise some pucks could be assigned to the
wrong team if they were still lying under a machine when the refbox was
started but then the pucks are moved to the opposite team.