protobuf_comm: sync with fawkes to support encryption
This brings in the new V2 framing protocol with encryption support. It
can be backwards compatible if desired (for peer connections). But this
is only temporary and will be removed at some point in the future.

For the LLSF at RoboCup 2014 the new protocol mustbe supported as per
decision of the Technical Committee.