llsf-fake-robot: adapt to new encrypted peer communication
The tool can now again announce a robot to the refbox and show received
information. If a team is assigned a name matching the one passed on the
command line, will check for encryption configuration and transparently
enable it. If the team name is unset or modified again (e.g. because the
refbox is restarted) encryption is disabled again, keeping communication
with the refbox up (unless a different team is setup with a different
encryption key).

This works on the fly and demonstrates how this can be implemented on
the side of the teams.