protobuf_comm: fix encrypted buffer size for multiple of block size
If the size of the message to be encrypted was a multiple of the block
size of the cipher later decryption would fail. The reason is that the
encrypted buffer size was calculated too small, hence the extra padding
block which must be added was missing, causing the encryption routine to
write the padding to unallocated memory. Then the message would be
resized to the proper length with only zeroes instead of proper padding.

Thanks to Frederik Zwilling for bringing this to my attention and
debugging patiently with me even though the world championship opening
game just started...