mps_com: Added tool to test mps
This tool is able to send modbuscommands to a MPS. There are following commands: Add a machine to the list, send or receive a command to a mps from the list, show the list. Each MPS type have his own functions that can be called. Its explained in the program what kind of functions you can call and how to use them.