rcll: automatic scoring on delivery reporting by color
When deliveries are reported by color of the delivered product, choose
the matching order most favorable to the delivering team and assign the
product to that order. Combined with the latest shell efforts this puts
more scoring knowledge into the refbox and less on the human referee.
Furthermore, this is a requirement for automatic scoring in simulation
and later for workpiece tracking deliveries.