rcll: award prod step points on late delivery, fix point calc
Award production points on late delivery as well.

The calculation of the grace period points was wrong and always yielded
the maximum number of points (the difference was calculated between the
current game time and the input fact game time, always being the same;
the proper way is to substract the delivery period end time from the
input game time).