rcll: account for down-time in preparation time, fixes #6
If a machine goes down, advance the time point since the machine may be
waiting for a workpiece to be fed. This keeps the time, the team has to
operate the machine while it is not down in the same, always the same.
This does extend the time the team has to feed a workpiece into a
machine, but circumvents issue #6.