rcll: directly assert product-delivered fact on delivery at DS
If the DS is in state PROCESSED, directly assert a product-delivered
fact with the new confirmed slot set to FALSE.

Previously, the fact was only created when a referee would enter a
delivery into the refbox. As we already know which order was delivered,
directly assert the fact to keep track of the order. This will require
further changes in the processing of the order.