rcll: use delivery time for scoring, not current game time
We store the time of the delivery when the DS goes into PROCESSED. This
is the actual time of the delivery, not the time when we receive the
delivery confirmation, which depends on the speed of the referee.
Therefore, use that time to compute the score.

Also rename the variable from ?game-time to ?delivery-time to make the
distinction clearer.