msgs: replace SetOrderDelivered msg by ConfirmDelivery msg
The new message confirms a delivery by the ID of the delivery, not by
order ID. It also contains a new field 'correct'. If it is set to true,
then the delivered product was as specified in the requeste order.
If it is set to false, then the wrong product was delivered.

The message format has changed completely, so rename the message and
also change the MSG_TYPE.