shell: only show unconfirmed deliveries in OrderDeliverMenu
This replaces a good part of the menu. We now only show deliveries that
have not been confirmed yet. We also no longer distinguish by order, but
by delivery ID.

This doesn't show the correct order configuration (i.e., the base, ring,
and cap colors) yet. It's also not yet possible to report an incorrect
delivery, a delivey is always assumed to be correct if confirmed.