shell: fix array access in OrderDeliverMenu
Accessing the order from the oinfo only works by index if we have all
orders in our list, which is not true anymore. Thus, instead of relying
on using the index as order ID, find the order that has the same ID as
the order ID in the current item.