rcll: handle asynchronous feedback from RS
Directly call mps-rs-mount-ring when the machine was prepared. This
assumes that the belt keeps running until it detects a workpiece in the
sensor, i.e., the belt possibly runs before a workpiece was put into the
machine.

After the ring was mounted, deliver it to the output of the RS and set
the machine to READY-AT-OUTPUT.