rcll: handle asynchronous feedback from CS
Directly call mps-cs-process when the machine was prepared. This assumes
that the belt keeps running until it detects a workpiece in the sensor,
i.e., the belt possibly runs before a workpiece was put into the
machine.

After the CS operation finished, deliver the workpiece to the output of
the RS and set the machine to READY-AT-OUTPUT.