refbox: add conveyor belt movements for CS operation
We also need to move the conveyor belt, just calling mount_cap() or
retrieve_cap() is not sufficient. First move the conveyor belt to the
mid position, send feedback to the refbox, and continue with the
operation. This way, the refbox knows that a workpiece has been put into
the machine.