rcll: separate machine timeout for a processing MPS
After a prepare message, only let a machine time out if it still
PREPARED, because we do not know exactly how long it needs to process.

Add a second timeout for a machine in state PROCESSING or PROCESSED. If
a machine times out while processing, it's not the robot's fault but a
problem with the MPS. In this case, only reset the machine without
setting it to broken.