rcll: Keep track of successfully PROCESSED production steps
Fact template (product-processed) tracks a single successfuly
PROCESSED operation done at an mps. Besides the process
info, it has the slots (workpiece ?id) (confirmed TRULE/FALSE)

The fact will be used at the central entity to determine scoring
of intermediate product.

Intended flow:
Initially, the product-processed NOT CONFIRMED, meaning
the fact is not linked to any workpiece. When workpiece
tracking is enabled, it will be confirmed upon AVAILBLITY
of the workpiece at the expected mps. Another way to
confirm the fact if workpiece tracking is disabled will
be by creating a workpiece when delivery is requested