rcll: Extend assign/resign oreder to  workpieces
A workpiece is assigned to any order that matches a perfect
 prefix of it colors.
A workpiece is resignd from an order whenever it becomes
 inconsistent with it.

The Flow of assigning and order to WP:
This means that the workpiece will be assigned a 'best guess'
order directly after it gets its base (ie, 'any' order that
has the same base). When a ring is mounted, the order will be
one step closer to the right guess and the old order will be
resigned as it becomes inconsistent with the recent operation.
(For Example, when starting to work on a C2 product with a reb
base at first the workpiece could be assigned to some C0 with
the same base color. As the workpiece gets its first ring, the
C0 order will be resignd off the workpiece and the correct C2
order will be assigned [see,assumption 1]. Furthermore, if the
team proceesds to mounting second ring of a wrong color on that
WP. The WP will be resigned from the C2 order)

This has some assumptions and grantees than needs to be implemented.

[assumption 1]
 A very specific assumption about the ring colors. All Cx
 compleuxities  have a unique ring color prefix. ie, the color
 of the first ring immediately determine the order of the workpiece.

[assumption 2]
 There is no harm in assigning the non inteded 'valid' order temporarly
 as long as the mistake is rectified as soon as its becomes clear.
 -For a Cx that happens by mounting the first ring.
 -For 2 C0's with a different cap that happens when mounting the cap.
 -For 2 identical products that happens when the delivery is requested.

[assumption 3]
 Any valid ambiguity should be corrected lastest at the delivery station
 when the bot requests the delivery of the order.

[grantee 1]
 The referee confirmation is ground truth. What the refree says the
 workpiece to look like overrides any 'wrong' decesions and rectifies
 any points given under 'wrong' order assigenment.

[Fact 1]
 In case of 2 identical orders (could happen with C0 for example).
 The only way to 'currently' know which one was intended is upon
 delivery.