rcll: Refactor 'product-delivered' fact, split into 2 facts
Product-delivered fact is removed and replaced by tow facts
representing its main functionalities. Namely, one to track the
info of the delivery requested at the DS, and another, to carry
the confirm requested from the referee. The splitting of
the confirmation and the book keeping has the advantage of
decoupling both functionalities and creates the needed freedom
to trigger the scoring only after the tracked workpiece has been
validated against the requested delivery.

Splitting into 'product-processed' fact and a 'referee-confirmation'
fact creates a generalized consistent way to ask for referee confirm
for a production step. This might be needed for exampled to confirm
a mounted cap on a workpiece that tracks a valid order and give score
for the cap step, even though the workpiece was wrong delivered.

Currently, the program flow is controlled by those facts as
follows:

[PRODUCT-PROCESSED]
* This fact tracks any production step done at any machine. Usually,
  when workpiece tracking is turned off, the only tracked step would
  be the one at the DS station for the delivery.
* When a production step is CONFIRMED (by which ever means!), the
  confirmation triggers awarding the scores of the corresponding step.
* How CONFIRMATION is done depends on the type of machine. For RS and
  BS, the process is confirmed when a workpiece becomes available at
  the machine. For DS and CS, confirmation is based on the referee
  decision.
* After the score for the step has been awarded. The fact field
  'scored' is set to TRUE. Score is only given once for a step.
* The fact also tracks for which order the step was 'believed' to
  produce. This allows us to reward score for intermediate steps
  which has not been processed more times than there is requested
  orders.
* This also allows us to reward points partially only for some of
  production step of WP. For example, if a team wishes to start over
  working on a C3 product because they lost the workpiece after
  mounting first ring, they shall only receive the points form the
  second ring on (in case the quantity required of that order is 1)

[referee-Confirmation]
* When a production step needs to be confirmed by the referee. This
  fact will track the status of the confirmation request.
* A referee-confirmation is always related to a product-processed fact
  that carries the step info to be confirmed.
* A confirmation status could be REQUIRED, CONFIRMED, DENIED. This allows
  us to distinguish between a negative confirmation, ie, between a Denied
  request and a yet non processed one. Previously, denied and invalid
  were both removed, only a positively confirmed one of yet unconfirmed
  were distinguished.