mps_comm: unify handling of modes mockup and simulation
Instead of having separate variables for the mockup and simulation,
introduce an enum ConnectionMode that can be either MOCKUP, SIMULATION,
or PLC, where MOCKUP means that we mock every machine interaction,
SIMULATION means that we connect to a simulated MPS, and PLC means that
we connect to a PLC on a real machine. Make the ConnectionMode a
constant member of each Machine, which is initialized during
construction.