rcll: get rid of the MPS state WAIT-IDLE
We no longer get feedback from the MPS in the way we used to, which also
makes the state WAIT-IDLE obsolete. Instead, we directly set the machine
to IDLE when get the signal that the machine is no longer READY (i.e.,
there is no workpiece in the output anymore).