rcll: get rid of machine slot proc-state
We no longer need the proc-state, as we directly use the task for
processing each step. The only case we still used the proc-state for was
the handling of a broken machine. To do this without the proc-state,
only reset the machine if there is a timeout and if the broken-since
time is not equal to 0.0. Additionally, set back the time to 0.0 when
resetting the machine.