rcll: switch to POST_GAME on finalize
If we are running a game and the refbox is terminated, switch to
POST_GAME before stopping. This prints a game summary and runs all other
post-game procedures. Also wait for POST_GAME before turning off all MPS
lights to avoid a race condition which would turn the light signal back
on, as the machine is supposed to have a different light signal if the
game is still running.