rcll: allow both RS slide events and protobuf add-base msg
Adapt the processing of the additional base on the RS slide such that we
process both the actual slide event and the protobuf message correctly,
no matter the order of events and messages. Do this by asserting an
intermittent mps-add-base-on-slide fact for the MPS, which is then
processed as usual. If we receive a protobuf message for the slide, also
assert that fact, but do not modify the base counter that is reported by
the MPS. This way, when the MPS reports the next base, that base will be
counted correctly, too.