rcll: remove redundant rule for wrong CS prepare msg
If we receive a prepare message to mount a cap if there is no cap in the
CS, we directly set the machine to BROKEN when we process the prepare
message. Thus, there is no need to check again whether we received the
wrong prepare message. Remove the redundant rule.