rcll: set machine to BROKEN on timeout while processing
The timeout may actually be caused by the team, e.g., if a robot
places the workpiece on the edge of the conveyor belt. Thus, handle this
like the other case, where no product is fed in time. If it is actually
a machine failure, the machine should still break, so the referee can
clear the machine.

Also use proc-start for the timeout instead of wait-for-product-since.
To do this, set proc-start wherever we set the state of the machine to
PROCESSING.