docker: add initial Dockerfile
The Dockerfile uses multi-stage builds to first build the refbox and
then to install the binaries and libraries into a separate image to
reduce the size of the final image. To avoid duplicating the required
packages, compute the dependencies from the binaries and libraries using
RPM's rpmdeps, which creates a list of dependencies for each file. This
list is then used to install all dependencies in the final image.