rcll: refactor game reset
The old way to reset the game did not work anymore, the game would be
reset repeatedly. Instead of relying on a game-reset fact, do everything
that's necessary for the reset in the function game-reset.

To avoid an issue with the order of file loading, allow resetting the
game by asserting a fact `(game-reset)`. This allows a game reset from
files where the function has not been defined yet.