rcll: No LOADED-WITH-MAX check needed on point deduction
Previously, in the old code, it used to be the case that the
bases-added slot was always updated in the machine fact by the
SLIDE COUNTER, keeping track of every base add operation, even
if the recently added base was above the capacity. The machine
would later break but the added-bases slot difference to the
used-bases will end up more the max capacity. Hence, it was needed
to make sure we do not deduct more points than we need.

This is no longer the case in the refactored code. The bases-added is
updated on an add event in the machine fact only if,the capacity isn't
violated. This makes sure that only usable bases are tracked.

Co-authored-by: Till Hofmann <hofmann@kbsg.rwth-aachen.de>