rcll: workpiece retrieved, IDLE does not mean retrieval
When a machine is IDLE, it does not mean that there is not workpiece
available at machine. (for example, if team placed the product on the
machine before preparing, triggering the Barcode reader while IDLE).

We can consider the workpiece retrieved when a machine completely
finished processing the product and before starting becoming available
again (WAIT-IDLE) or when the machine BREAKS.