rcll: reset cs-retrieved only on BROKEN recovery
Resting the slot before breaking the CS is useless (if not unreasonable).
Breaking recovery, should take care of resetting machine status information
in all related slots before it goes back to idle.

The information contained in the status slots before it BROKEN might be
useful to assess the reason of Breaking or to make decision before
recovery.

Co-authored-by: Till Hofmann <hofmann@kbsg.rwth-aachen.de>