rcll: remove check for identical colors but different order
The rule was to make sure to change the order ID if the
confirmed delivery operation related to a workpiece that
tracks another order with the same colors.

This is no longer needed.
Currently, the workpiece is verified on delivery confirmation, ensuring
that, a confirmed delivery operation implies that its workpiece looks like
and tracks the correct order the referee had confirmed.