rcll: warn if assumption of unique cx-order prefixes violated
Assumption:
Assigning orders to workpieces has a very specific assumption about
the ring colors. We assume that the color of the first ring fully
determines the complexity class. Furthermore  we assume that there
is only a single order of complexity >0 each. Hence, knowing the first
ring will immediately determine the order of the workpiece."