rcll: ensure first 4 Cx-orders will have unique prefix
An order complexity determines what is the first ring-color.

Yet, According to the assumption needed for workpiece to order
assignment (look rule workpiece-assign-order). There is also only a
single order of each complexity. Currently, this is not the case and
we have tow orders of type C1. This could cause some error in tracking.

This fix ensures that the unused ring-color will be used as a first ring
for orders of higher complexity that has more than a single order.