rcll: implement the simplified storage station with a single product
This SS only has a single C0 on its conveyor belt, which is moved to the
output when receiving a prepare for a RECEIVE. There is no need to
specify a storage slot. The machine will break if prepared for a STORE
operation, as the current SS is not able to do that.