rcll: wait 5 seconds during storage station processing
If we directly dispense the workpiece, we may be too fast for a team to
get the state of the storage station, which may result in a broken SS,
as the team keeps preparing the station. Thus, wait 5 seconds before
actually dispensing the workpiece. This also resembles the actual SS
more closely, as it takes some time to retrieve the workpiece from the
shelf.