rcll: give full production steps in failsafe for late delivery
We should not assume that the individual production steps happened at
the same time as the delivery, as this results in no points given if the
delivery is late. Instead, if we are not tracking, simply assume that
all steps happened at time 0.1 to guarantee that they are before the
delivery deadline. This restores the old behavior of giving full points
for production steps, even if the delivery is late.