rcll: always give intermediate points even if done late
Points for ring and cap mounting should always be awarded regardless if
they were performed later than the order delivery window.
This restores the old behavior of giving full points for production
steps, even if the delivery is late.

This is an alternative implementation to extend the solution implemented
at PR #44.

This fixes issue #42.

Co-authored-by: Till Hofmann <hofmann@kbsg.rwth-aachen.de>