rcll: only set MPS to READY-AT-OUTPUT if workpiece has been seen
We sometimes miss workpieces and then disable the tracking completely
for the whole game. Instead of doing this, do not set the machine to
READY-AT-OUTPUT if it did not see a workpiece and the tracking is
enabled. This allows the referee and or the machine to recover, either
by manually moving the workpiece into the barcode scanner, or by moving
the belt.