rcll: wait if MPS goes DOWN while PREPARED but not PROCESSING
A machine may go DOWN when it is prepared but no product has been feed
into the machine yet. In this case, also delay the timeout for the
product, so a team can still feed the product into the machine after it
comes back.

Also unconditionally delay the `proc-start` time to simplify the code.