rcll: only randomize ring colors when generating a new field
Do not randomize the ring color assignment unconditionally anymore.
Instead, only randomize the assignment if a new playing field is
generated.

This also removes some ties between the ring color assignment and the
ring colors of the orders. Previously, the first ring of the first C1
and first C2 orders would always be on the first RS, while the first
ring of the first C3 would always be on the second RS. This is no longer
the case, machine assignment and order assignment are now completely
independent.