mps_comm: restructure command queue for OpcUaMachines
Instead of putting a std::function in the queue, make a queue of
instructions to send and call the class method `send_instruction()` to
actually send the instruction to the machine. This separates the
queue and the instruction details more cleanly and also avoids a bug
where the instructions would be processed in the wrong order if sending
the instruction failed the first time.