mps_comm: fix off-by-one error in MPSRegister enum handling
The MPSRegister STATUS_READY_BASIC was never used because its value was
used as the size of the registers and thus never included when looping
over the array. Add a pseudo register LAST, which can be used to loop
over the enum, and which allows for later changes to the enum.