mps_comm: skip queue when sending identify instruction
We cannot put the identify instruction into the queue, because it needs to
be sent before we send any other instruction. Instead, skip the queue
and directly send the instruction.

Also implement it in the base class OpcUaMachine. We already have the
machine type and there is no need to implement this function in each
sub-class.