cfg: add config to skip payment checks for ring stations
When a MPS is inoperable, games can still be played normally by
simulating machine actions on refbox side. However, a problem occurs in
case a ring station has to be simulated due to missing feedback when a
base is added to the slide. Since no machine instruction is sent by the
acting robots in that case, it is impossible for the refbox to simulate
a slide counter increment in real-time.

To fix this, the refbox can simply bypass the slide counter checks for
sufficient payment and instead increment the counter to the desired
amount then. This allows to keep using the inoperable MPS with human
assistance in the following form:

1. manually mounting rings if the whole machine is inoperable
2. keeping track of actual successful payments and manually breaking the
   simulated RS in case
   a) more than 3 payments were done without consumption
   b) insufficiently many bases were present when a ring mount was
      requested
3. awarding the points for payed bases that were not consumed
   when the game ends.

As a first step towards this objective, create a configurable list of
ring stations that should not get checked for sufficient base payments.