rcll: make sure simulated slide puts provide correct score
If multiple slide puts are asserted subsequently it may happen that the
game time did not progress in between. This leads to the point fact not
getting asserted as they are identical for multiple slide puts. To
prevent this, simply make sure that time elapsed since the last asserted
points.