tools: extend rcll-refbox-instruct to wait for a given number of robots
Sometimes, we may want to wait for robots, e.g., when running in
simulation. Instead of guessing how long the robots need to connect,
extend the rcll-refbox-instruct tool to wait for some given number of
robots. This also works in conjunction with the other wait options,
e.g., waiting for a game phase.