mps-comm: scale individual station wait times
All the individual machine processing steps may be scaled as well.
However, in order to avoid mps state transitions that happen to fast to
be observable, always have a minimum duration of 1 second.

The delivery station is a special case that gets a minimal duration of
two seconds as it consists of fewer mps states.