rcll: adapt to latest storage station iteration
the latest iteration of the storage station can be controlled by
specifying the storage position in terms of a shelf (between 0 and 7) and a
shelf slot (between 0 and 5).

The storage station control is still work in progress and this is just
an early iteration aiming to get a mockup mps running as soon as
possible.