rcll: add SS relocation + SS related costs
In order to make use of all shelf slot from the SS special care has to
be taken when accessing slots that could be blocked by other products
that are positioned in between the SS gripper and the target position.

Each SS shelf has the following format:

1 3 5 7
0 2 4 6

 L_|
  |

Hence uneven indices might be inaccessible if the corresponding front
slots are occupied.
The Refbox has to relocate blocking workpieces to other accessible
positions before using a blocked slot in the back.
There are 3 cases:
1. Relocation is required, a free slot is accessible -> relocate to free
   slot
2. Relocation is required, no free slot is accessible, there are empty
   slots left -> first make a slot accessible by relocating a product to
   the back, then relocate the blocking product to the newly accessible
   before proceeding
3. Relocation is required, no free slot is accesible, there are no other
   free slots in the back available -> machine breaks
   So if all slots on all shelves are occupied, back slots cannot be
   retrieved.
   Also, if only a single slot in the back is free, then this slot
   cannot be used to store a product.