rcll: refactor instruction-msg processing to functions
The original rule was way too long and logic changes before the
individual machine processing steps lead to long confusing diffs if
the indentation of the case-distinction was effected.

Since this branch already faced this issue we might as well fix it now
as the diffs are a mess now anyways.