websocket: use environment variable SHAREDIR to locate schamas
The schemas are shared required at runtime, hence they should be addressed
by an environment variable that is known at runtime regardless of where
the refbox is installed. EXEC_SHAREDIR is the variable that is also used
for protobuf messages and clips files, hence we can use it for this use
case as well.