mps_placing_clips: check connectivity of the generated field
The MPS placement should not prevent any free zone from being reachable.
This can be ensured by encoding the field as a grid graph while removing
all edges to fields that are occupied.
If the vertices of the free zones form a single connected component,
then all free zones are orthogonally connected.