rcll: make sure start-time is always set
The start-time is usually set once the PRE_GAME phase ends.
If for some reason the refbox skips PRE_GAME (e.g. by being manually
instructed via the helper scripts), then this might not be the case.
Simply make sure to set the start-time whenever it is not set yet.