rcll: fix bugs with writing machine-history
As a safety measure explicitly check if ?temp-fact exists before
retracting it.
Also, if does not exist this means that the machine history fact is
identically with the current machine fact. But in that case there was a
bug that the bson doc for that fact was not build, which is now fixed as
well.