mps_comm: connect to OPC-UA machines in the background
This allows us to complete the refbox initialization without having
connected to all machines. This is helpful especially in simulation,
where the refbox first needs to announce the machine positions before
the simulation spawns the machines. In this case, the refbox cannot
connect to the machine at initialization, but the machines are not set
up until the refbox has been initialized completely, leading to a
deadlock.