mps_comm: do not load configuration again, pass it instead
With the recent addition of customized configs there is no main
configuration file by default. While that file can be explicitly
generated, that should not be necessary just because the mps_comm
component wants to use a config value.

Instead, pass the parsed configuration to the machine factory, which can
delegate any config value to specific machines.