msgs: rework Machine Report messages to deal with markerless reports
When reporting machines that do not have AR tags teams can only report a
machine type and zone at first.
They need to get feedback about the actual name and team color in return
(if the report was correct), which then can be used to properly report
the specific machine (or its mirrored counterpart in case a machine of
the opposite team was encountered).