rcll: handle incoming machine reports without set machine name
If a team wants to report a machine that does not have an AR tag, then
it can only report a machine type and not its name.
Therefore, the refbox needs to answer by providing the name and the team
of the reported mps in case an incoming report has set the type + a zone
instead of the name.
This is done using MachineTypeFeedback messages.
Those only contain the name + zone, if an mps of the matching type is
actually in that zone and if a team did not try to report that machine
type to often already.
That means that for stations that only occur once (BS, SS, DS) a team
can only report these types once in hope to get an answer.
Reporting it again (e.g., at a different zone) leads to the report being
ignored.
Similarly, each team has two attempts to report RS and CS types.