rcll: properly rectify confirmed delivered workpiece
Whenever a delevery is confirmed, the associated workpiece should be
matching the belonging order.
Ideally, there should be past operations on the respective machines
needed to assemble the product.
If not, create the missing steps as the delivery was confirmed.
However, this case should not happen and it might be useful to create
rules for that situation in the future.