mps_placing_clips: fix wrong logic on blocked DS zones
The "input_only_mps" set consists of machines, which only have an
operable input, but no output, which are currently only delivery
stations. The Problem here is, that the angles were confused as the
the input side blocks the bottom zone only in angles higher than 180
degree:

         225°     270°    315°
            O      O     O
         [ ]      [ ]     [ ]
        I          I         I
Bottom    ^        ^       ^

Vice versa:

          45°      90°    135°
Top       v        v       v
            I      I     I
         [ ]      [ ]     [ ]
        O          O         O

Where [ ] denotes a machine, I (O) is the position of the input (output) side and Top/ Bottom point to the respective zone of the DS that should be blocked in this part of the code.