rcll: simplify order scoring and improve delivery deadline penalties
Firstly, move the points for mounting the last ring on a product to the
delivery. This simplifies the logic when rewarding points for ring
mounts.
Secondly, rework the punishment for late deliveries to be more
impactful.
Starting from the delivery end, more points are deduducted based on the
elapsed time in N=4 steps. These are spaced evenly based on the orginal
delivery interval, e.g., if the delivery window ist [60,180], then every 30
seconds ( (180-60)/4 ) the point deduction gets more severe.

The amount deducted depends on a fraction of the rewarded delivery points T:
floor(75% * T * X/N), where X is the number of thresholds that are crossed.
25% of the delivery points cannot be lost, no matter how late the delivery.

In the example above, if the product is a C0 with 20 delivery points,
then delivering at 181s yields a penalty of -3 (75% of 20 = 15, 1/4 of
15 = 3.75, rounded down).
delivering at 211 would instead yield -7, delivering at 241 results in
-11 and a delivery any time after 270 seconds yields -15 points.