rcll: let DS wait when delivering early
It makes no sense to receive 0 points on early deliveries. A better
model would be that the early delivery blocks the delivery station as it
cannot be processed fully, yet.
Therefore, the DS simply waits in state PREPARED until the delivery time
approaches when instructed to deliver a product too early.
It cannot be used in the meantime (unless of course it is manually
broken).
Once the delivery window starts, full points are scored as expected.