mongodb: fix wrong errors when loading field layouts from mongodb
The problem is the machine_meta array in the game report document.
It contains fact information for different templates (all the different
meta facts with unique templates per machine type).

The previous option to load arrays into facts could not handle this
properly, as it complained that there were no clips fact found for those
documents that did not belong to the requested fact template.

To fix this, distinguish between the loading of all or just some facts
from document arrays.
If all facts should be loaded, then a failure occurs if there is a
document that cannot be matched to a fact in clips.
If only some facts should be loaded, then failure occurs only if not a
single document matches an existing fact.