rcll: refactor time out of gamestate
The time being part of the gamestate was annoying, because it forces
developers to mute the facts in the debug log.
Also, rules that only want to act in the correct phase and only want to
fire based on fact changes did not work. A good example are the rules
around websocket messages. These typically want to update the websocket
if a fact changes (and sometimes, if additionally the phase is correct).

before this change this actually lead to flooding the websocket with
updates.