websocket: catch crashes on connect and print errors
I had a case where during a running game the refbox crashed for no
apparent reason. There was a frontend running, but nothing was done with
it.
At some point i checked the refbox and it crashed due to a failed client
connection attempt. Not sure what was going on there, but this should
fix the issue.