rcll: sync globals for game with confval facts
The current system is awkward. Not only are there two different config
paths to describe field specs (for challenges and for regular settings),
the globals are also only written once (when the initial config is
loaded).
To patch up issues regarding different config paths, make sure that
both confval facts update according to the global value.
That way, confval facts for the challenges and for the game agree with
eachother.
The Second issue still persists but will be a topic of a separate PR.
Idally, the whole config system should be reworked so that clips confval
facts represent the actual true used config, not the yaml files they are
initialized from or the globals they are partially synced to.